+++ src/stb/stb_image_resize.h
@@ -636,7 +636,7 @@ static float stbir__srgb_to_linear(float f)
     if (f <= 0.04045f)
         return f / 12.92f;
     else
-        return (float)pow((f + 0.055f) / 1.055f, 2.4f);
+        return (float)powf((f + 0.055f) / 1.055f, 2.4f);
 }
 
 static float stbir__linear_to_srgb(float f)
@@ -644,7 +644,7 @@ static float stbir__linear_to_srgb(float f)
     if (f <= 0.0031308f)
         return f * 12.92f;
     else
-        return 1.055f * (float)pow(f, 1 / 2.4f) - 0.055f;
+        return 1.055f * (float)powf(f, 1 / 2.4f) - 0.055f;
 }
 
 #ifndef STBIR_NON_IEEE_FLOAT
@@ -762,7 +762,7 @@ static float stbir__filter_trapezoid(float x, float scale)
     float t = 0.5f + halfscale;
     STBIR_ASSERT(scale <= 1);
 
-    x = (float)fabs(x);
+    x = (float)fabsf(x);
 
     if (x >= t)
         return 0;
@@ -786,7 +786,7 @@ static float stbir__filter_triangle(float x, float s)
 {
     STBIR__UNUSED_PARAM(s);
 
-    x = (float)fabs(x);
+    x = (float)fabsf(x);
 
     if (x <= 1.0f)
         return 1 - x;
@@ -798,7 +798,7 @@ static float stbir__filter_cubic(float x, float s)
 {
     STBIR__UNUSED_PARAM(s);
 
-    x = (float)fabs(x);
+    x = (float)fabsf(x);
 
     if (x < 1.0f)
         return (4 + x*x*(3*x - 6))/6;
@@ -812,7 +812,7 @@ static float stbir__filter_catmullrom(float x, float s)
 {
     STBIR__UNUSED_PARAM(s);
 
-    x = (float)fabs(x);
+    x = (float)fabsf(x);
 
     if (x < 1.0f)
         return 1 - x*x*(2.5f - 1.5f*x);
@@ -826,7 +826,7 @@ static float stbir__filter_mitchell(float x, float s)
 {
     STBIR__UNUSED_PARAM(s);
 
-    x = (float)fabs(x);
+    x = (float)fabsf(x);
 
     if (x < 1.0f)
         return (16 + x*x*(21 * x - 36))/18;
@@ -886,9 +886,9 @@ static int stbir__get_filter_pixel_width(stbir_filter filter, float scale)
     STBIR_ASSERT(filter < STBIR__ARRAY_SIZE(stbir__filter_info_table));
 
     if (stbir__use_upsampling(scale))
-        return (int)ceil(stbir__filter_info_table[filter].support(1/scale) * 2);
+        return (int)ceilf(stbir__filter_info_table[filter].support(1/scale) * 2);
     else
-        return (int)ceil(stbir__filter_info_table[filter].support(scale) * 2 / scale);
+        return (int)ceilf(stbir__filter_info_table[filter].support(scale) * 2 / scale);
 }
 
 // This is how much to expand buffers to account for filters seeking outside
@@ -901,9 +901,9 @@ static int stbir__get_filter_pixel_margin(stbir_filter filter, float scale)
 static int stbir__get_coefficient_width(stbir_filter filter, float scale)
 {
     if (stbir__use_upsampling(scale))
-        return (int)ceil(stbir__filter_info_table[filter].support(1 / scale) * 2);
+        return (int)ceilf(stbir__filter_info_table[filter].support(1 / scale) * 2);
     else
-        return (int)ceil(stbir__filter_info_table[filter].support(scale) * 2);
+        return (int)ceilf(stbir__filter_info_table[filter].support(scale) * 2);
 }
 
 static int stbir__get_contributors(float scale, stbir_filter filter, int input_size, int output_size)
@@ -1016,8 +1016,8 @@ static void stbir__calculate_sample_range_upsample(int n, float out_filter_radiu
     float in_pixel_influence_upperbound = (out_pixel_influence_upperbound + out_shift) / scale_ratio;
 
     *in_center_of_out = (out_pixel_center + out_shift) / scale_ratio;
-    *in_first_pixel = (int)(floor(in_pixel_influence_lowerbound + 0.5f));
-    *in_last_pixel = (int)(floor(in_pixel_influence_upperbound - 0.5f));
+    *in_first_pixel = (int)(floorf(in_pixel_influence_lowerbound + 0.5f));
+    *in_last_pixel = (int)(floorf(in_pixel_influence_upperbound - 0.5f));
 }
 
 // What output pixels does this input pixel contribute to?
@@ -1031,8 +1031,8 @@ static void stbir__calculate_sample_range_downsample(int n, float in_pixels_radi
     float out_pixel_influence_upperbound = in_pixel_influence_upperbound * scale_ratio - out_shift;
 
     *out_center_of_in = in_pixel_center * scale_ratio - out_shift;
-    *out_first_pixel = (int)(floor(out_pixel_influence_lowerbound + 0.5f));
-    *out_last_pixel = (int)(floor(out_pixel_influence_upperbound - 0.5f));
+    *out_first_pixel = (int)(floorf(out_pixel_influence_lowerbound + 0.5f));
+    *out_last_pixel = (int)(floorf(out_pixel_influence_upperbound - 0.5f));
 }
 
 static void stbir__calculate_coefficients_upsample(stbir_filter filter, float scale, int in_first_pixel, int in_last_pixel, float in_center_of_out, stbir__contributors* contributor, float* coefficient_group)
@@ -1041,7 +1041,7 @@ static void stbir__calculate_coefficients_upsample(stbir_filter filter, float sc
     float total_filter = 0;
     float filter_scale;
 
-    STBIR_ASSERT(in_last_pixel - in_first_pixel <= (int)ceil(stbir__filter_info_table[filter].support(1/scale) * 2)); // Taken directly from stbir__get_coefficient_width() which we can't call because we don't know if we're horizontal or vertical.
+    STBIR_ASSERT(in_last_pixel - in_first_pixel <= (int)ceilf(stbir__filter_info_table[filter].support(1/scale) * 2)); // Taken directly from stbir__get_coefficient_width() which we can't call because we don't know if we're horizontal or vertical.
 
     contributor->n0 = in_first_pixel;
     contributor->n1 = in_last_pixel;
@@ -1093,7 +1093,7 @@ static void stbir__calculate_coefficients_downsample(stbir_filter filter, float 
 {
     int i;
 
-    STBIR_ASSERT(out_last_pixel - out_first_pixel <= (int)ceil(stbir__filter_info_table[filter].support(scale_ratio) * 2)); // Taken directly from stbir__get_coefficient_width() which we can't call because we don't know if we're horizontal or vertical.
+    STBIR_ASSERT(out_last_pixel - out_first_pixel <= (int)ceilf(stbir__filter_info_table[filter].support(scale_ratio) * 2)); // Taken directly from stbir__get_coefficient_width() which we can't call because we don't know if we're horizontal or vertical.
 
     contributor->n0 = out_first_pixel;
     contributor->n1 = out_last_pixel;
